feat: criação do componente card para facilitar a manutenção
