+++ screens/Drawer/DrawerNavigator.js
@@ -4,8 +4,6 @@ import HomeScreen from '../home/HomeScreen';
 import { View, Text } from 'react-native';
 import CustomDrawer from './CustomDrawer';
 import ProfileScreen from '../profile/ProfileScreen';
-import CreateCard from '../CreateCard/CreateCard';
-import App from '../ReadCard/ReadCardEspacada.js';
 
 const Drawer = createDrawerNavigator();
