+++ screens/ReadCard/ReadCard.js
@@ -1,17 +1,16 @@
-import React, { useRef, useState, useEffect } from 'react';
-import { Animated, Dimensions, StyleSheet, Text, TouchableWithoutFeedback, View, Alert, TouchableOpacity } from 'react-native';
+import React, { useState, useEffect } from 'react';
+import { View, Dimensions, StyleSheet, Text, Alert, TouchableOpacity } from 'react-native';
 import TopBar from '../../components/TopBar/TopBar';
+import CardComponent from '../../components/Card/CardComponent';
 import IconTextButton from '../../components/IconTextButton/IconTextButton';
 import ButtonImage from '../../components/ButtonImage/ButtonImage';
-import Ionicons from '@expo/vector-icons/Ionicons';
 
 //Tem que passar o baralho para entrar nessa tela.
 export default function App({ navigation }) {
   const [baralho, setBaralho] = useState('Redes');
-  const [pergunta, setPergunta] = useState('');
-  const [resposta, setResposta] = useState('');
   const [cards, setCards] = useState([]);
   const [currentIndex, setCurrentIndex] = useState(0);
+  const [cardRef, setCardRef] = useState(null);
 
   useEffect(() => {
     fetch('http://localhost:3001/cards')
@@ -21,50 +20,6 @@ export default function App({ navigation }) {
       })
       .catch(err => console.error('Erro ao buscar cards:', err));
   }, []);
-
-  const [isFlipped, setIsFlipped] = useState(false);
-  const flipAnimation = useRef(new Animated.Value(0)).current
-
-  const frontInterpolate = flipAnimation.interpolate({
-    inputRange: [0, 180],
-    outputRange: ['0deg', '180deg'],
-  });
-
-  const flipToFrontStyle = {
-    transform: [{ rotateY: frontInterpolate }]
-  };
-
-
-  const backInterpolate = flipAnimation.interpolate({
-    inputRange: [0, 180],
-    outputRange: ['180deg', '360deg'],
-  });
-
-  const flipToBackStyle = {
-    transform: [{ rotateY: backInterpolate }]
-  };
-
-
-  const flipCard = () => {
-    if (isFlipped) {
-
-      Animated.spring(flipAnimation, {
-        toValue: 0,
-        friction: 8,
-        tension: 10,
-        useNativeDriver: true,
-      }).start();
-    } else {
-
-      Animated.spring(flipAnimation, {
-        toValue: 180,
-        friction: 8,
-        tension: 10,
-        useNativeDriver: true,
-      }).start();
-    }
-    setIsFlipped(!isFlipped);
-  };
 
   function handleDificuldade(cardId, dificuldade) {
     fetch(`http://localhost:3001/cards/${cardId}`, {
@@ -115,10 +70,8 @@ export default function App({ navigation }) {
     );
   }
 
-
   return (
     <View style={styles.container}>
-      {/* Precisa incrementar o topbar igual a tela home (Leticia) */}
       <TopBar
         image1={require('../../assets/backIcon.png')}
         onPress1={() => navigation.navigate('Home')}
@@ -128,24 +81,12 @@ export default function App({ navigation }) {
         style2={styles.image}
       />
 
-
       <View style={styles.content}>
-
-        <View style={styles.showCard}>
-          <TouchableWithoutFeedback onPress={flipCard}>
-            <View style={styles.cardContainer}>
-              {/*PERGUNTA*/}
-              <Animated.View style={[styles.front, styles.card, flipToFrontStyle]}>
-                <Text style={styles.text}>{cards[currentIndex]?.pergunta || 'Sem perguntas'}</Text>
-              </Animated.View>
-
-              {/*RESPOSTA*/}
-              <Animated.View style={[styles.back, styles.card, flipToBackStyle]}>
-                <Text style={styles.text}>{cards[currentIndex]?.resposta || 'Sem resposta'}</Text>
-              </Animated.View>
-            </View>
-          </TouchableWithoutFeedback>
-        </View>
+        <CardComponent 
+          pergunta={cards[currentIndex]?.pergunta}
+          resposta={cards[currentIndex]?.resposta}
+          ref={setCardRef}
+        />
 
         <View style={styles.answer}>
           <View style={styles.informationTitle}>
@@ -160,7 +101,6 @@ export default function App({ navigation }) {
             <TouchableOpacity style={styles.answerButton}
               onPress={() => {
                 handleDificuldade(cards[currentIndex].id, 1);
-                if (isFlipped) flipCard();
               }}
             >
               <Text style={styles.answerTexts}>Difícil</Text>
@@ -169,19 +109,13 @@ export default function App({ navigation }) {
             <TouchableOpacity style={styles.answerButton}
               onPress={() => {
                 handleDificuldade(cards[currentIndex].id, 0);
-                if (isFlipped) flipCard();
-              }
-              }
+              }}
             >
               <Text style={styles.answerTexts}>Fácil</Text>
             </TouchableOpacity>
           </View>
         </View>
-
       </View>
-
-
-
     </View>
   );
 }
@@ -197,28 +131,12 @@ const styles = StyleSheet.create({
     width: 50,
     height: 50,
   },
-  showCard: {
-    alignItems: 'center',
-    marginTop: 25,
-  },
-  cardContainer: {
-    width: width - 50,
-    height: height / 3,
-  },
-  front: {
-    backgroundColor: '#E2C2FB',
-  },
-  back: {
-    backgroundColor: '#96D289',
-  },
   card: {
     width: width - 50,
     height: height / 3,
     alignItems: 'center',
     justifyContent: 'center',
     borderRadius: 10,
-    position: 'absolute',
-    backfaceVisibility: 'hidden',
   },
   text: {
     fontSize: 20
